requirements.txt
asgiref==3.4.1
colorama==0.4.4
Django==3.2.8
django-extensions==3.1.5
django-jazzmin==2.4.9
django-ninja==0.16.1
pretty-errors==1.2.25
pydantic==1.8.2
pytz==2021.3
sqlparse==0.4.2
typing==3.7.4.3
typing_extensions==4.0.1
tzdata==2021.5

Pillow~=9.0.0
python-jose==3.3.0